Move Executable Trait under Node namespace
